[sailfish-connman-iptables-plugin] Added a bash script as test tool for the iptables D-Bus API.
The test can be run only on the device. Requires connman and sailfish-connman-iptables-plugin to be running.

Adds rules to iptables using the D-Bus API, checks that rule exists in iptables and removes the rule from iptables. Executes one add and remove operation for each interface.

Quits if one test fails and returns 1. Otherwise 0 if all tests are success.