[code] Implement ICMP rule adding and removal support. JB#42674
This commit adds support for adding ICMP rules to iptables. A new method
was added: "RuleICMP". This function has following attributes:
 - string: table name
 - string: chain name
 - string: target name
 - string: source IP address [optional, can be empty]
 - string: destination IP address [optional, can be empty]
 - uint16: ICMP type
 - uint16: ICMP code
 - uint16: operation [ADD = 0, REMOVE = 1]

For IPv4 ICMP types are 0...43 and codes 0...15 according to:
https://www.iana.org/assignments/icmp-parameters/icmp-parameters.xhtml

For IPv6 ICMP types are 0...161 and codes 0...255 according to:
https://www.iana.org/assignments/icmpv6-parameters/icmpv6-parameters.xhtml

If either of the parameters (type or code) is invalid error INVALID_ICMP
(15) is returned. Otherwise the behavior in error situations is
identical to the other Rule* methods.

For each add and removal of ICMP rule the signals are emitted:
 - RuleICMPAdded
 - RuleICMPRemoved