[test] Change nemo to defaultuser and fix ConnMan path in e-t-e. JB#61353
There is no nemo user on the devices anymore and defaultuser is used.
Changed this in the end-to-end tests and made further changes on this
easier. Dropped nemo alltogether from the test scripts.

Changed to use the proper location /home/.system/var/lib/connman for the
iptables IPv4 saves. This is used by saving end-to-end tests to check if
the changes were saved correctly.