[test] Use numeric protocols for rule verification. JB#61352
Newer iptables does use only numeric protocols when listing (-L) is done
with -n flag. Previously, at least up to 1.8.7, the protocols were used
as strings. This change caused end-to-end tests to fail and is fixed here.

Changed also the IP address to be something that might not be in use on
the testing network (hopefully), 192.168.10. seems to be used for some
pre-setup networks.

Also cleanup tabs in the test.